refactor: 변수 이름, 패키지, 예외 메시지, 테스트 이름 수정
