+++ src/main/java/racingcar/controller/RacingCarGame.java
@@ -1,6 +1,7 @@
 package racingcar.controller;
 
 import racingcar.domain.*;
+import racingcar.domain.car.Cars;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
