+++ src/main/java/racingcar/domain/car/Car.java
@@ -1,4 +1,6 @@
-package racingcar.domain;
+package racingcar.domain.car;
+
+import racingcar.domain.GoOrStop;
 
 public class Car {
     private static final int MAX_NAME_LENGTH = 5;
@@ -16,13 +18,13 @@ public class Car {
 
     public static void validateEmptyName(String input) throws IllegalArgumentException {
         if (input.isEmpty()) {
-            throw new IllegalArgumentException();
+            throw new IllegalArgumentException("입력값은 빈 문자열일 수 없습니다.");
         }
     }
 
     public static void validateNameLength(String input) {
         if (isOverLength(input)) {
-            throw new IllegalArgumentException();
+            throw new IllegalArgumentException("자동차의 이름은 5자이하만 가능합니다.");
         }
     }
 
@@ -49,6 +51,6 @@ public class Car {
         result.append(name);
         result.append(" : ");
         result.append(location);
-        return String.valueOf(result);
+        return result.toString();
     }
 }
