+++ src/test/java/racingcar/controller/RacingCarGameTest.java
@@ -3,7 +3,4 @@ package racingcar.controller;
 import org.junit.jupiter.api.Test;
 
 public class RacingCarGameTest {
-    @Test
-    void 쉼표를_기준으로_문자열을_분리() {
-    }
 }
