+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,13 +1,14 @@
 package racingcar.domain;
 
 import org.junit.jupiter.api.Test;
+import racingcar.domain.car.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class CarTest {
     @Test
-    void 자동차_이름이_빈_문자열인지_검증() {
+    void 자동차_이름이_빈_문자열인_경우_예외_처리() {
         String input = "";
         assertThatThrownBy(() -> {
             Car.validateEmptyName(input);
@@ -15,7 +16,7 @@ public class CarTest {
     }
 
     @Test
-    void 자동차_이름이_5자초과인_경우_검증_실패() {
+    void 자동차_이름이_5자초과인_경우_예외_처리() {
         String input = "두강두강무늬무늬";
         assertThatThrownBy(() -> {
             Car.validateNameLength(input);
