+++ src/test/java/racingcar/domain/RoundTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class RoundTest {
     @Test
-    void 시도할_회수가_숫자가_아닌_경우_검증_실패() {
+    void 시도할_회수가_숫자가_아닌_경우_예외_처리() {
         String input = "a";
         assertThatThrownBy(() -> {
             Round.validateIsNumber(input);
@@ -14,7 +14,7 @@ public class RoundTest {
     }
 
     @Test
-    void 시도할_회수가_음수가_아닌_경우_검증_실패() {
+    void 시도할_회수가_음수가_아닌_경우_예외_처리() {
         int input = -1;
         assertThatThrownBy(() -> {
             Round.validateIsNegativeNumber(input);
